pyproject.toml
[tool.poetry]
name = "discord-twitter-link-embedder"
version = "0.1.0"
description = ""
authors = ["agis09"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord = "^2.3.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
